Add test for pmaps bug
The function s12df_to_s12l intented to convert evt_max events from a
DataFrame to a dictionary failed to stop when evt_max index is missing
in the input DataFrame. This bug is fixed and some functionality
added:

+ Passing a negative number as evt_max results in a “read all” behaviour.
+ Default value for evt_max is None (which is converted to -1 for the Cython function).
+ evt_max is implemented as long int to avoid precision limitations.

+++ invisible_cities/core/pmaps_functions_c.pxd
@@ -6,6 +6,6 @@ cimport numpy as np
 import numpy as np
 
 
-cpdef cdf_to_dict(int df_index, int evt_max,
+cpdef cdf_to_dict(int df_index, long int evt_max,
                   int [:] df_event, int [:] df_peak,
                   float [:] df_time, float [:] df_ene)
